Feat: Hiding anime toolbar when fullscreening
I wanted to get a result similar to the Preview Mac app. However, I don’t know if there is a way to do that… I’ve searched for a long time so if anyone knows a way, please reach me through a GitHub issue or @zhich on Discord.

With help from this post: https://stackoverflow.com/questions/76366878/how-do-i-hide-my-toolbar-in-swiftui-when-full-screened